CORE-18698: move FilteredTransaction implementtion, factory, tests to ledger libs
