+++ libs/ledger/ledger-common-data/src/main/java/net/corda/ledger/common/data/transaction/filtered/factory/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.ledger.common.data.transaction.filtered.factory;
+
+import org.osgi.annotation.bundle.Export;
