+++ libs/ledger/ledger-common-data/src/main/kotlin/net/corda/ledger/common/data/transaction/filtered/ComponentGroupFilterParameters.kt
@@ -1,6 +1,5 @@
-package net.corda.ledger.common.flow.transaction.filtered.factory
+package net.corda.ledger.common.data.transaction.filtered
 
-import net.corda.ledger.common.data.transaction.filtered.FilteredTransaction
 import net.corda.v5.crypto.merkle.MerkleProof
 import net.corda.v5.crypto.merkle.MerkleProofType
 import java.util.function.Predicate
